+++ src/deisa/dask/constants.py
@@ -1,0 +1,7 @@
+from typing import Final
+
+FEEDBACK_QUEUE_PREFIX: Final[str] = "deisa_feedback_queue_"
+CALLBACK_PREFIX: Final[str] = "deisa_cb_"
+DEFAULT_SLIDING_WINDOW_SIZE: int = 1
+CLIENT_KEY: Final[str] = "deisa"
+KEY_PREFIX: Final[str] = "deisa"